Create ServerSocketListener
