+++ src/main/java/echoServer/Listener.java
@@ -1,5 +1,7 @@
 package echoServer;
 
+import java.io.IOException;
+
 public interface Listener {
-    Connection listen();
+    Connection listen() throws IOException;
 }
