+++ src/test/java/echoServer/SocketConnectionTest.java
@@ -2,12 +2,7 @@ package echoServer;
 
 import org.junit.jupiter.api.Test;
 
-import java.io.ByteArrayInputStream;
-import java.io.ByteArrayOutputStream;
 import java.io.IOException;
-import java.io.InputStream;
-import java.io.OutputStream;
-import java.net.Socket;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
@@ -34,28 +29,4 @@ public class SocketConnectionTest {
         assertEquals("Hello\n", output);
     }
 
-    private class TestSocket extends Socket {
-
-        private final String inputString;
-        private ByteArrayOutputStream outputStream;
-
-        public TestSocket(String inputString) {
-            this.inputString = inputString;
-        }
-
-        @Override
-        public OutputStream getOutputStream() throws IOException {
-            outputStream = new ByteArrayOutputStream();
-            return outputStream;
-        }
-
-        public String output() {
-            return outputStream.toString();
-        }
-
-        @Override
-        public InputStream getInputStream() throws IOException {
-            return new ByteArrayInputStream(inputString.getBytes());
-        }
-    }
 }
